v8 - add MHFU (USA) codes, label some not-working MHFU (EU) codes
IncognitoMan ported some of his MHP2G codes to MHFU (USA) (ULUS-10391) - these have been added. Unfortunately, the MHP2G codes don't work as-is on MHFU (EU) (my version of the game). They have been labelled as such. Notably, the "HUD Off" code (my favourite!) has an additional check on USA/EU versions of the game, so for now it only works on JP. I need to test some of the USA codes on the EU version to see if they work.